pyproject.toml
[project]
name = "discord-tracker"
version = "0.1.0"
description = "A bot for tracking various Discord events."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord>=2.3.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.0",
]
